Ignore bad iframe generation script
We reload the iframe with a new src attribute. The script they use to
generate it appears to, for some reason, strip out the 'www.' before
'auburn.edu', even if it's there; this causes the HTTPS request to fail,
since our security cert is only valid on *.auburn.edu.

As such, the new code grabs the URL that *should* be there and supplants
the bad one generated (just adds in a 'www.').